Fixed unit tests. Proguard was shrinking out functions it needed. Also added more UI tests.
