+++ app/proguard-rules.pro
@@ -1,5 +1,5 @@
-# Some methods are only called from tests, so make sure the shrinker keeps them.
--keep class com.example.android.architecture.blueprints.** { *; }
+# Keep classes contianing methods used in the testing modules
+-keep class us.spencer.habittracker.** { *; }
 
 -keep class android.support.v4.widget.DrawerLayout { *; }
 -keep class android.support.test.espresso.IdlingResource { *; }
