+++ app/src/main/res/layout/add_habit_frag.xml
@@ -59,6 +59,7 @@
                 android:id="@+id/add_habit_freq_spinner"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:layout_marginLeft="16dp"
                 android:layout_marginStart="16dp"
                 android:layout_marginTop="8dp"
                 app:layout_constraintStart_toEndOf="@+id/add_habit_freq_tv"
